+++ src/components/ogre/widgets/QuickHelp.cpp
@@ -55,7 +55,7 @@ std::list<HelpMessage>::const_iterator QuickHelp::messagePosition(const HelpMess
 	}
 
 	//Does the Id exist? If it does, return the position.
-	int location = 1;
+	size_t location = 1;
 	for (auto list_iterator = mTutorialText.begin(); list_iterator != mTutorialText.end(); ++list_iterator) {
 
 		if (list_iterator->mId == message.mId) {
